Add first version of Eldar colorscheme
This colorscheme is a heavily modified version of Elflord designed for
dark backgrounds. It displays that same in the terminal and the GUI
(because it is designed for the terminal).

It makes use of Al Bunden's ColourAssignment map and unpacking function,
originally from: http://www.cgtk.co.uk/vim-scripts/bandit