Add Background and Text color variables
This allows customization of the background and text colors. In the
terminal the background and text default to the terminal colors, but may
be overridden with `g:eldar_term_background` and `g:eldar_term_text` In
the GUI the background and text default to "Black" and "White"
respectively. They may be overridden with `g:eldar_background` and
`g:eldar_text`.